0206 modify density&main, 2 addline funcs work great

+++ week1/densityMap.cpp
@@ -132,9 +132,9 @@ void DensityMap::addLine(glm::vec3 p1, glm::vec3 p2, std::vector<float> vals) {
 std::vector<float> DensityMap::getVertices() {
 	std::vector<float> vertices;
 
-	for (int i = 0; i < dim - 1; i++) {
-		for (int j = 0; j < dim - 1; j++) {
-			for (int k = 0; k < dim; k++) {
+	for (float i = 0; i < dim - 1; i++) {
+		for (float j = 0; j < dim - 1; j++) {
+			for (float k = 0; k < dim; k++) {
 				float v1[3] = { i, j, k };
 				float v2[3] = { i + 1, j, k };
 				float v3[3] = { i, j + 1, k };
@@ -151,9 +151,9 @@ std::vector<float> DensityMap::getVertices() {
 		}
 	}
 
-	for (int i = 0; i < dim - 1; i++) {
-		for (int j = 0; j < dim; j++) {
-			for (int k = 0; k < dim - 1; k++) {
+	for (float i = 0; i < dim - 1; i++) {
+		for (float j = 0; j < dim; j++) {
+			for (float k = 0; k < dim - 1; k++) {
 				float v1[3] = { i,j, k };
 				float v2[3] = { i + 1, j, k };
 				float v3[3] = { i, j, k + 1 };
@@ -170,9 +170,9 @@ std::vector<float> DensityMap::getVertices() {
 		}
 	}
 
-	for (int i = 0; i < dim; i++) {
-		for (int j = 0; j < dim - 1; j++) {
-			for (int k = 0; k < dim - 1; k++) {
+	for (float i = 0; i < dim; i++) {
+		for (float j = 0; j < dim - 1; j++) {
+			for (float k = 0; k < dim - 1; k++) {
 				float v1[3] = { i, j, k };
 				float v2[3] = { i, j + 1, k };
 				float v3[3] = { i, j, k + 1 };
